+++ CHANGES.md
@@ -9,6 +9,7 @@
 * refactored `GitHub_API::get_repo_meta()` to use more efficient API call, gets forks also, thanks @egifford
 * introduce some variability to transient expiration per plugin
 * switch to storing repo data in options table instead of using transients
+* fixed branch switching with extended naming [#520](https://github.com/afragen/github-updater/issues/520), thanks @joelworsham
 
 #### 6.1.1 / 2016-11-29
 * hotfix to flush cache during upgrade routine